ADDED: the darkness dependency update
